requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-enumfields==2.1.1
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
PyJWT==2.9.0
pytz==2024.1
sqlparse==0.5.1
tzdata==2024.1
